version 3.07
What's new in release 3.07
Release date: 10 March 2010

  * Fast Synchronous Data Mover Facility (Guy Desbiens)
  * Diagnose 210, 250, 260 (Harold Grovesteen)
  * Extended Diagnose 204 feature (Jan Jaeger)
  * Complete Diagnose 24 (Harold Grovesteen)
  * Configuration-Topology Facility (Fish)
  * HFP-Unnormalized-Extensions Facility (Harold Grovesteen)
  * CMPSC performance improvements (Bernard van der Helm)
  * uptime command (Fish)
  * Raise XPNDSIZE limit to 1048576MB (Roger Bowler)
  * MAXCPU and LPARNUM configuration statements (Roger Bowler)
  * Add capacity model identifiers to MODEL config statement (Roger Bowler)
  * SCLPROOT configuration statement (Jan Jaeger)
  * Add "noclear" option to printer and card punch devices (Jay Maynard)
  * Socket printer support (Fish)
  * 3705 SNA device support (Max H. Parke)
  * TTY and 2741 support for 2703 (Max H. Parke)
  * Tracing enhancements (Jan Jaeger)
  * Allow configure --enable-external-gui for Unix builds
  * Enable tun/tap emulation for 64-bit Windows builds (Ivan Warren)
  * 64-bit Windows support (Roger Bowler)
  * Raise MAX_CPU_ENGINES limit to 64 (Roger Bowler, Ivan Warren)
  * Numerous bug fixes (Ivan Warren, Fish, Jan Jaeger, Bernard van der Helm,
Roger Bowler, Jay Maynard)

+++ conspawn.c
@@ -11,9 +11,9 @@
 /*       invoking this program.                                      */
 /*-------------------------------------------------------------------*/
 
-// $Id: conspawn.c,v 1.3 2006/12/08 09:43:18 jj Exp $
+// $Id: conspawn.c 5147 2009-02-01 21:27:16Z rbowler $
 //
-// $Log: conspawn.c,v $
+// $Log$
 // Revision 1.3  2006/12/08 09:43:18  jj
 // Add CVS message log
 //
@@ -109,7 +109,7 @@ int main(int argc, char* argv[])
         else
             p = NULL;
 
-        rc = (int) ShellExecute( hwnd, lpOperation, lpFile, lpParameters, lpDirectory, nShowCmd );
+        rc = (intptr_t) ShellExecute( hwnd, lpOperation, lpFile, lpParameters, lpDirectory, nShowCmd );
 
         if (p)
             free(p);
